moved marked down to /config/markdown and added html files to static folder
