+++ static/daenerys-targaryen.html
@@ -1,0 +1,80 @@
+<p><strong>Meta Title:</strong> Why Daenerys Targaryen will win the Game of Thrones</p>
+<p><strong>Meta Description:</strong> Daenerys is playing to win and there’s no one in the Seven Kingdoms who can stop
+    her now. Learn all about her on got show!</p>
+<hr/>
+<p>
+    <img src="http://vignette1.wikia.nocookie.net/gameofthrones/images/5/58/Daenerys-Targaryen.jpg/revision/latest?cb=20141202204150&amp;path-prefix=de"/>
+    &gt; Daenerys Targaryen. Game of Thrones‘ mother of dragons.</p>
+<h3 id="h1-why-daenerys-targaryen-will-win-the-game-of-thrones">H1: Why Daenerys Targaryen will win the Game of
+    Thrones</h3>
+<p>Daenerys. Khaleesi. Mother of Dragons. Mhysa. There’s quite a few names for this woman – and even more reasons to be
+    afraid of her. Ever since her father King Aerys II Targaryen was ousted from the Iron Throne during Robert
+    Baratheon’s rebellion, she’s been desperate to take it back – and with it the control over the seven kingdoms. Let’s
+    face it: The moment Daenerys gave her OK for Khal Drogo to kill her (admittedly annoying) brother, we knew that girl
+    played to win. Here’s why we’ll see here sitting on the Iron Throne soon.</p>
+<h4 id="h2-shes-commanding-the-unsullied">H2: She’s commanding the Unsullied</h4>
+<p><img src="https://musingsofalifelongnerd.files.wordpress.com/2015/05/unsullied1.jpg"/> &gt; Alt text: Men in black:
+    The Unsullied in Game of Thrones</p>
+<p>In case you want to take a city back, it is always a good idea to bring a private army. Daenerys’ has apparently done
+    her homework on warmonger best practices and bought her very own battalion, the Unsullied, by burning their former
+    owner alive (#justdaenerysthings). Even with significantly less testicles or nipples compared to your average Essos
+    man, those guys are anything and everything you could wish for when you absolutely need some people dead. Still in
+    doubt? <a href="https://www.youtube.com/watch?v=0TB3rgJlJdo&amp;nohtml5=False">See them in action.</a></p>
+<h4 id="h2-daenerys-is-the-peoples-woman">H2: Daenerys is the people’s woman</h4>
+<p><img src="http://8840-presscdn-0-18.pagely.netdna-cdn.com/wp-content/uploads/2014/04/danythrone1.jpg"/> &gt; Alt
+    text: Always watching out for the little man: Daenerys Targaryen</p>
+<p>With the US presidential elections approaching, maybe it’s time for a small thought experiment. Imagine you could
+    vote for… &gt; Please display the images in a row next to each other here</p>
+<div class="figure">
+    <img src="http://rack.1.mshcdn.com/media/ZgkyMDE1LzA2LzA1LzQwL1RoZV9IaWdoX1NwLmE3OTExLmpwZwpwCXRodW1iCTEyMDB4OTYwMD4/fbe4acb5/9d5/The_High_Sparrow_promo_pic-640x426.jpg"/>
+</div>
+<blockquote>
+    <p>Alt text: The High Sparrow in Game of Thrones</p>
+</blockquote>
+<p>
+    <img src="http://vignette1.wikia.nocookie.net/gameofthrones/images/7/78/Tommen_Season_6.jpg/revision/latest/scale-to-width-down/270?cb=20160211212139"/>
+    &gt; Alt text: Tommen Baratheon in Game of Thrones</p>
+<p>
+    <img src="http://vignette4.wikia.nocookie.net/gameofthrones/images/4/41/Dany_Mother_Mhysa.png/revision/latest?cb=20130610044643"/>
+    &gt; Alt text: The people’s woman: Daenerys Targaryen in Game of Thrones</p>
+<ol style="list-style-type: lower-alpha">
+    <li><p>A religious fanatic in charge of what is basically the Game of Thrones equivalent of ISIS</p></li>
+    <li><p>A wimpy teenager with no experience whatsoever who should have stayed a minor character anyway</p></li>
+    <li><p>A young, dynamic civil rights activist, running around Essos freeing slaves in the thousands, who’s known for
+        her zero-tolerance policy regarding the establishment</p></li>
+</ol>
+<p>… Who would you choose? Heh? Not so hard, right? Daenerys Targaryen is pretty much our Game of Thrones Bernie
+    Sanders. Only in slightly more attractive. Sorry Bern.</p>
+<blockquote>
+    <p>Placeholder: Twitter Sentiment Analysis Graph for Daenerys Targaryen</p>
+</blockquote>
+<blockquote>
+    <p>Alt text: See what Twitter thinks about Daenerys Targaryen with GoT show's sentiment analysis tool</p>
+</blockquote>
+<p>As you can see, she’s even among the most popular characters <a href="link%20to%20ranking">(rank 4, 10th of
+    April)</a> in the real world! Feel the Daern, people!</p>
+<h4 id="h2-she-has-a-bunch-of-dragons">H2: She has a bunch of dragons</h4>
+<p><img src="https://blurppy.files.wordpress.com/2015/01/game-thrones-season-4-pictures.jpg"/> &gt; Daenery’s pet
+    dragon. Makes your dog look like a hamster in comparison.</p>
+<p>… Is there a reason we’re still debating this? Seriously? Even though her lathery scaled friends started out as cute
+    little hatchlings in season 1, those things are terrifying now. We all saw them eat the leaders of the great
+    families in the season 5 episode “Kill the boy” (S05E05) – and as countless fantasy franchises suggest, if you have
+    dragons by your side, chances are you’re going to bring home a victory.</p>
+<h5 id="h3-sooo-things-are-set-right">H3: … Sooo, things are set, right?</h5>
+<p>Well, almost. In Game of Thrones, there’s always a hint of uncertainty – and you have to admit that a lot of people
+    want to see her dead. <a href="https://www.youtube.com/watch?v=MIGYQr6ldJk">Take this assassination attempt on
+        youtube for example.</a> We in the got.show editorial department are confident that she’ll end up being
+    victorious, even though our machine learning algorithm is a lot less optimistic. The algorithm mostly takes family
+    history into account, and since her’s was cleared off the map, her calculated likelihood of death is pretty high.
+    See our latest prediction here: &gt; Placeholder, Daenerys Targaryen PLOD</p>
+<blockquote>
+    <p>Alt text: Daenery’s calculated percentage likelihood of death</p>
+</blockquote>
+<p>No reason to be worried, though. Valar morghulis, remember?</p>
+<h5 id="h3-follow-daenerys-long-journey-through-essos">H3: Follow Daenerys long journey through Essos</h5>
+<blockquote>
+    <p>Placeholder: our awesome GoT Map about Daenerys</p>
+</blockquote>
+<blockquote>
+    <p>Alt Text: Daenerys Targaryen put on the map – Game of Thrones character movements</p>
+</blockquote>
