+++ static/house-stark.html
@@ -1,0 +1,74 @@
+<p><strong>Meta Title:</strong> What’s next for Bran, Arya &amp; Sansa? House Stark in Game of Thrones</p>
+<p><strong>Meta Description:</strong> House Stark has been struggling season after season. Will they make it through the
+    next? Find out at GoT show!</p>
+<hr/>
+<p><img src="http://awoiaf.westeros.org/images/thumb/5/51/House_Stark.PNG/250px-House_Stark.PNG"/> &gt; Alt text: House
+    Stark of Winterfell: Coat of arms ### H1: Will Bran, Arya &amp; Sansa be the last of their kind in Game of Thrones?
+</p>
+<p>House Stark of Winterfell has been having kind of a rough patch lately, don’t you think so? Although “rough patch” is
+    quite the understatement when you’re talking about multiple fatalities in the family. And “lately” refers to the
+    very first season of Game of Thrones, where we all taught an impressive lesson on how little George RR. Martin cares
+    about his main characters.</p>
+<p>It seems like winter’s been coming and staying for Bran, Arya and Sansa. But what exactly have they been up to lately
+    – and where are they headed?</p>
+<h4 id="h2-bran-stark">H2: Bran Stark</h4>
+<p>
+    <img src="http://vignette3.wikia.nocookie.net/gameofthrones/images/e/e9/Bran_Stark_Season_Six.jpg/revision/latest/scale-to-width-down/270?cb=20160220011003"/>
+    &gt; Alt text: Bran Stark, Game of Thrones</p>
+<p>It’s been awfully quiet around Bran this year, since HBO decided to keep him from us during Season 5. While we can
+    only hope someone got fired for that, we sure as hell are excited for his appearance in april! Remember season 4’s
+    epic finale, where the young stark boy was promised <a href="https://www.youtube.com/watch?v=7YXnI3HgZbc">the
+        ability to fly after meeting the three-eyed raven?</a> Seems like his buddy Hodor didn’t carry him around all
+    this time for nothing after all. Or, how he likes to put it:</p>
+<p>
+    <img src="http://vignette2.wikia.nocookie.net/gameofthrones/images/5/52/Profil-Hodor-1.jpg/revision/latest/scale-to-width-down/250?cb=20130902151852&amp;path-prefix=de"/>
+    &gt; Alt text: Hodor, hodor hodor. From Game of Thrones. Hodor.</p>
+<p>Hodor hodor, hodor. Hodor.</p>
+<p>Did you know that Hodor is actually the second most popular Game of Thrones character on Twitter? See for
+    yourself!</p>
+<blockquote>
+    <p>Placeholder: Twitter Tool Hodor</p>
+</blockquote>
+<blockquote>
+    <p>Alt text: What people think about Hodor on Twitter. Game of Thrones Sentiment Analysis.</p>
+</blockquote>
+<p><a href="link%20to%20popularity%20ranking">See the full list here</a></p>
+<h4 id="h2-sansa-stark">H2: Sansa Stark</h4>
+<p>
+    <img src="http://vignette1.wikia.nocookie.net/gameofthrones/images/5/5b/Sansa_Season_Six_Snow.jpg/revision/latest/scale-to-width-down/270?cb=20160213183617"/>
+    &gt; Alt text: Sansa Stark, Game of Thrones</p>
+<p>Oh boy. Where do I start? Sansa’s been having a troubled family life ever since her dad passed away. After two
+    unhappy marriages she has fled her husband yet again – with Theon Greyjoy by her side. Quite an unusual combination,
+    after all they’ve been through, but I guess nothing unites people quite as well as a mutual disgust for Ramsay
+    Bolton. … That is of course, if they are still alive. Which is hard to say at this point – unless you calculate the
+    likelihood with a machine learning algorithm, like the fancy one we’ve implemented right here on this page! Always
+    looking out for fellow Game of Thrones fans here. Currently, we estimate Sansa’s survival chances to be:</p>
+<blockquote>
+    <p>Placeholder: PLOD Sansa Stark</p>
+</blockquote>
+<blockquote>
+    <p>Alt text: Sansa Stark’s survival chances in Game of Thrones season 6</p>
+</blockquote>
+<h4 id="h2-arya-stark">H2: Arya Stark</h4>
+<div class="figure">
+    <img src="http://awoiaf.westeros.org/images/thumb/a/a0/Arya_Stark_by_AniaEm.jpg/250px-Arya_Stark_by_AniaEm.jpg"/>
+</div>
+<p>We’re not the only ones who’ve been kept dark when it comes to the Game of Thrones universe for the past few months.
+    For Arya Stark, season 6 is going to look a lot like 50 Shades of Grey since she got punished with blindness by the
+    Many-Faced-God for taking a life that wasn’t hers to take. I guess those things happen when you surround yourself
+    with that kind of people. No reason to give us that look Arya, it’s certainly not our fault!</p>
+<p>
+    <img src="http://vignette4.wikia.nocookie.net/gameofthrones/images/9/9e/Arya_Season_6.jpg/revision/latest/scale-to-width-down/270?cb=20160211211717"/>
+    &gt; Alt text: Arya Stark, Game of Thrones</p>
+<h3 id="h3-wait-a-minute">H3: Wait a minute…</h3>
+<p>… didn’t we forget somebody here? Yes, we did! Jon Snow, the other children’s half-brother! <s>Oh, but he’s dead
+    anyway</s> Wrong! He’s as fresh as a daisy. How we know? Machine learning. <a href="urltojonsnowlandingpage">See for
+    yourself!</a></p>
+<h3 id="h3-follow-the-starks-around-their-journey-through-westereos">H3: Follow the Stark’s around their journey through
+    Westereos</h3>
+<blockquote>
+    <p>Placeholder: GoT Map with Arya, Bran &amp; Sansa</p>
+</blockquote>
+<blockquote>
+    <p>Alt text: Arya, Bran &amp; Sansa Stark put on the Game of Thrones map.</p>
+</blockquote>
