+++ static/jon-snow-is-not-dead.html
@@ -1,0 +1,67 @@
+<p><strong>[Meta Title]</strong> Why Jon Snow is not dead - Game of Thrones <strong>[Meta Description]</strong> We at
+    got.show set up our own machine learning algorithm to predict Jon Snow's fate and came to surprising conclusions!
+    <strong>Optimized for Keywords:</strong> Jon Snow, Game of Thrones, got</p>
+<hr/>
+<h1 id="h1-why-jon-snow-is-not-dead">H1: Why Jon Snow is not dead</h1>
+<p><a href="https://www.youtube.com/watch?v=IxI8aPISq8I">“The past is already written, the ink is dry”</a> – so says the
+    teaser for the sixth season of the popular TV show Game of Thrones. And until recently, that’s what we at <strong>got.show</strong>
+    believed too. 9 long months after we saw our favourite character <strong>Jon Snow</strong> join his father up in GoT
+    heaven in the season 5 finale “Mother’s Mercy” however, we have reason to believe that he’s going to stick around on
+    earth with the rest of us for a little longer. You don’t think so? Well, then <strong>you know nothing, Jon
+        Snow.</strong></p>
+<h2 id="h2-4-reason-jon-snow-is-still-alive">H2: 4 reason Jon Snow is still alive</h2>
+<h3 id="h3-our-very-own-machine-learning-algorithm-suggests-it.">H3: Our very own machine learning algorithm suggests
+    it.</h3>
+<p>We at <strong>got.show</strong> are firm believers that artificial intelligence &amp; machine learning can have a
+    more beneficiary impact on humanity than <a
+            href="http://www.theverge.com/2016/3/15/11213518/alphago-deepmind-go-match-5-result">embarrassing us at our
+        own games</a> – that’s why we <strong>created an algorithm that calculates Game of Thrones characters’
+        likeliness of death</strong>! Luckily, <strong>Jon is only at 11.6%</strong>, which is overall pretty low.
+    You’ll be glad to know that <a href="https://en.wikipedia.org/wiki/Ramsay_Bolton">Ramsay Bolton</a> for example is
+    at over 60%! Furthermore, being Lord Commander of the Night's Watch makes him one of the safest guys in all of
+    Westeros, since it is the title with the lowest death rate in the series. <a
+            href="https://got-stats.herokuapp.com/ranking">Read more about our awesome Life &amp; Death Predictions
+        here!</a> &gt; Placeholder 1) Infographic Image <span class="citation">@togiberlin</span> &gt; Culture, title
+    &amp; gender play a huge role in terms of survival chances on Games of Thrones.</p>
+<h3 id="h3-he-is-way-too-popular-to-be-killed-off.">H3: He is way too popular to be killed off.</h3>
+<p>I know, I know – being popular couldn’t save a lot of our other favourite Game of Thrones characters, but with Jon
+    it’s different. Just see what an outrage the season 5 finale caused on Twitter with our Twitter Sentiment Analysis
+    App – and then think about the gigantic plot twist of bringing him back to the living! &gt; Placeholder 2) Twitter
+    graph. &gt; How does Twitter feel about Jon Snow? See for yourself with our Game of Thrones Sentiment analysis.</p>
+<h3 id="h3-kit-harington-jon-snows-actor-has-been-on-set-for-the-sixth-season-game-of-thrones">H3: Kit Harington (Jon
+    Snow’s actor) has been on set for the sixth season Game of Thrones!</h3>
+<p>
+    <a href="http://www.independent.co.uk/arts-entertainment/game-of-thrones-season-6-not-only-is-jon-snow-alive-but-he-s-wearing-stark-armour-a6669801.html">Read
+        about it in the independent.</a>. And if that’s still not enough for you, what do you think about this official
+    HBO poster? <img src="http://www.hbo.com/custom-assets/img/series/game-of-thrones/season-6-tease.jpg"/> &gt; Alt
+    Text: Official Announcement poster for the sixth season Game of Thrones, featuring Jon Snow.</p>
+<h3 id="h3-he-just-cant-be-dead.">H3: He just can’t be dead.</h3>
+<p>I can’t even. Who’s going to care for Sam &amp; Gilly? Who will handle the wildlings? Let me tell you one thing: If
+    Leonardo di Caprio can drag himself through the freezing cold for weeks after <a
+            href="https://en.wikipedia.org/wiki/The_Revenant_%282015_film%29">fighting with a goddamn grizzly bear</a>,
+    why shouldn’t Kit Harington survive a minor stabbing, heh?</p>
+<p><strong>Still not convinced?</strong> Then you’ll be glad to hear that the season 6 premiere of Game of Thrones is
+    going to air on the 24th April 2016 already - <a href="http://www.daystillgameofthrones.com/">soon we will know.</a>
+</p>
+<h2 id="h2-follow-jons-last-latest-steps-here">H2: <strong>Follow Jon’s <s>last</s> latest steps here!</strong></h2>
+<blockquote>
+    <p>Placeholder 3) GOT Map Alt Text: Follow Jon Snow’s footsteps with our Game of Thrones map! Shortly after leading
+        the wildlings to Castle Black to save them from the nearing White Walkers, Jon had to learn about Stannis
+        Baratheon’s fate from Melisandre. Then the unthinkable happened. Where will we see him next? Stay tuned at
+        got.show</p>
+</blockquote>
+<hr/>
+<h1 id="about-the-placeholders">About the Placeholders</h1>
+<h2 id="placeholder-1-infographic-about-plod-contributing-features">Placeholder 1) Infographic about PLOD contributing
+    features</h2>
+<p><span class="citation">@togiberlin</span> Please make an infographic referencing the stats from this <a
+        href="https://github.com/Rostlab/JS16_ProjectB_Group6/issues/60#issuecomment-201949253">ProjectB issue</a>. I’d
+    only focus on Culture, Title &amp; Gender info. Maybe emphasize Jon Snow’s survival change explicitly again when it
+    comes to the Lord Commander of the Night’s Watch. Project F will be able to give you some guidelines on size &amp;
+    formatting while they’re arranging the template page.</p>
+<h2 id="placeholder-2-twitter-sentiment">Placeholder 2) Twitter sentiment</h2>
+<p>Here we’ll just serve a view from ProjectD_Group4’s Twitter graph (my group). I’ve added startDate &amp; endDate
+    parameters to the function that creates the graph, so you can specify the default date Range when loading the tool.
+    I’d go for something around May 20 – June 20 2015 (season finale aired on June 14th)</p>
+<h2 id="placeholder-3-got-map">Placeholder 3) GOT Map</h2>
+<p>Just our amazing Map with Jon Snow added as a character. I’d show all the available data.</p>
